Day 6 Part one completed

+++ day-6/src/main.rs
@@ -6,7 +6,7 @@ fn part_one() -> io::Result<()> {
     let reader = io::BufReader::new(file);
 
     let mut lines_iter = reader.lines();
-    let mut result: u16 = 1;
+    let mut result: u32 = 1;
 
     if let Some(Ok(line1)) = lines_iter.next() {
         let content_time: Vec<&str> = line1.split(':').collect();
@@ -18,11 +18,11 @@ fn part_one() -> io::Result<()> {
             let content_distance: Vec<&str> = line2.split(':').collect();
             let distances: Vec<&str> = content_distance.get(1).unwrap().split_whitespace().collect();
             println!("Distances: {:?}", distances);
-            let mut ways: u16 = 0;
+            let mut ways: u32 = 0;
             for (index, &time_str) in times.iter().enumerate() {
-                if let Ok(time) = time_str.parse::<u16>() {
+                if let Ok(time) = time_str.parse::<u32>() {
                     if let Some(record_distance_str) = distances.get(index) {
-                        if let Ok(record_distance) = record_distance_str.parse::<u16>() {
+                        if let Ok(record_distance) = record_distance_str.parse::<u32>() {
                             for holding_time in 0..=time {
                                 println!("Holding time {}", holding_time);
                                 let new_record = holding_time * (time - holding_time);
